Added edit and delete for existing entries.
Editing opens "Edit movie" window, and loads existing information.
Saving this edit will cause the original to disappear and adds it in the bottom of the list.

Delete ("x") will remove the movie from the GUI and the intern list.

Fixed not updating entry count in the GUI (bottom left).

No saving.
No searching.
No sorting.