Alpha 1.1.0 work in progress OMDb support
With Alpha 1.1.0 comes the OMDb support.
It is now possible to send search requests to OMDb on creating new entries, this will autofill all available information.
If there is only one fitting movie it will autofill.
If there are more than one an selection window will pop up (this is not formatted yet).

Changes
- Added OMDb key save/load [if no key present will ask for one]
- Added OMDb key reset button
- Changed button '+' to 'new entry'
- Added 'search' button on new entry
- Added search on OMDb
- Added model structures for json responses from OMDb
- Added alerts on search [no title / too many results / movie not found]
- Added dependencies httpclient (apache) and jackson-databind
- Moved Movie and MovieList to package model
- Moves Main and GUI in package app
- Added movie selection window [shown when search will return more than 1 entry] (no format)
- Added autofill for OMDb requests [direct for 1 movie in response and from selection window]